moved five modules into one big folder
