+++ bpaxos/common_testpub.go
