+++ bpaxos/consensus/consensus_test.go
@@ -2,12 +2,16 @@ package main
 
 import (
 	"testing"
-	"All-On-Cloud-9/proposer/node"
+	"All-On-Cloud-9/bpaxos/consensus/node"
 	"All-On-Cloud-9/common"
 )
 
 func TestHandleReceive(t *testing.T) {
 	t.Errorf("STUB: Test for HandleReceive still unimplemented")
+}
+
+func TestReachConsensus(t *testing.T) {
+	t.Errorf("STUB: Test for ReachConsensus still unimplemented")
 }
 
 func TestSendResult(t *testing.T) {
