+++ bpaxos/consensus/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"All-On-Cloud-9/dependency/node"
+	"All-On-Cloud-9/bpaxos/consensus/node"
 	"All-On-Cloud-9/common"
 	"github.com/nats-io/nats.go"
 	"fmt"
@@ -9,27 +9,26 @@ import (
 )
 
 func main() {
-	dependency_node := depsnode.NewDepsServiceNode()
-	// dependency_node.Stub()
-	go func(dep_node *depsnode.DepsServiceNode) {
+	cons := consensus.ConsensusServiceNode{}
+	go func(cons *consensus.ConsensusServiceNode) {
 		socket := common.Socket{}
 		_ = socket.Connect(nats.DefaultURL)
-		socket.Subscribe(common.LeaderToDeps, func(m *nats.Msg) {
-			fmt.Println("Received leader to deps")
+		socket.Subscribe(common.ProposerToConsensus, func(m *nats.Msg) {
+			fmt.Println("Received proposer to consensus")
 			data := common.MessageEvent{}
 			json.Unmarshal(m.Data, &data)
-			newMessage := dep_node.HandleReceive(&data)
+			newMessage := cons.HandleReceive(&data)
 			sentMessage, err := json.Marshal(&newMessage)
 
 			if err == nil {
-				fmt.Println("deps can publish a message to leader")
-				socket.Publish(common.DepsToLeader, sentMessage)
+				fmt.Println("consensus can publish a message to proposer")
+				socket.Publish(common.ConsensusToProposer, sentMessage)
 			} else {
 				fmt.Println("json marshal failed")
 				fmt.Println(err.Error())
 			}
 
 		})
-	}(&dependency_node)
+	}(&cons)
 	common.HandleInterrupt()
 }
