+++ bpaxos/leader/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"All-On-Cloud-9/leader/node"
+	"All-On-Cloud-9/bpaxos/leader/node"
 	"All-On-Cloud-9/common"
 	"github.com/nats-io/nats.go"
 	"fmt"
