+++ bpaxos/proposer/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"All-On-Cloud-9/proposer/node"
+	"All-On-Cloud-9/bpaxos/proposer/node"
 	"All-On-Cloud-9/common"
 	"github.com/nats-io/nats.go"
 	"fmt"
