+++ bpaxos/proposer/proposer_test.go
@@ -2,20 +2,12 @@ package main
 
 import (
 	"testing"
-	"All-On-Cloud-9/replica/node"
+	"All-On-Cloud-9/bpaxos/proposer/node"
 	"All-On-Cloud-9/common"
 )
 
 func TestHandleReceive(t *testing.T) {
 	t.Errorf("STUB: Test for HandleReceive still unimplemented")
-}
-
-func TestAddDepsToGraph(t *testing.T) {
-	t.Errorf("STUB: Test for AddDepsToGraph still unimplemented")
-}
-
-func TestExecVertices(t *testing.T) {
-	t.Errorf("STUB: Test for ExecVertices still unimplemented")
 }
 
 func TestSendResult(t *testing.T) {
