+++ bpaxos/replica/node/replica.go
