+++ bpaxos/replica/replica_test.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"testing"
-	"All-On-Cloud-9/consensus/node"
+	"All-On-Cloud-9/bpaxos/replica/node"
 	"All-On-Cloud-9/common"
 )
 
@@ -10,8 +10,12 @@ func TestHandleReceive(t *testing.T) {
 	t.Errorf("STUB: Test for HandleReceive still unimplemented")
 }
 
-func TestReachConsensus(t *testing.T) {
-	t.Errorf("STUB: Test for ReachConsensus still unimplemented")
+func TestAddDepsToGraph(t *testing.T) {
+	t.Errorf("STUB: Test for AddDepsToGraph still unimplemented")
+}
+
+func TestExecVertices(t *testing.T) {
+	t.Errorf("STUB: Test for ExecVertices still unimplemented")
 }
 
 func TestSendResult(t *testing.T) {
